+++ app/scripts/com/2fdevs/videogular/directives/videogular.js
@@ -20,6 +20,10 @@
  *
  * **This parameter is disabled in mobile devices** because user must click on content to prevent consuming mobile data plans.
  *
+ * @param {boolean} [vgStartTime=-1] vgStartTime Number value or a String with a scope name variable to start playing the video at a certain time.
+ *
+ * @param {boolean} [vgVirtualClipDuration=-1] vgVirtualClipDuration Number value or a String with a scope name variable for a length to limit the video playback to.
+ *
  * @param {object} vgCuePoints Bindable object containing a list of timelines with cue points objects. A timeline is an array of objects with the following properties:
  * - `timeLapse` is an object with two properties `start` and `end` representing in seconds the period for this cue points.
  * - `onEnter` callback called when user enters on a cue point. callback(currentTime, timeLapse, params)
@@ -34,6 +38,8 @@
    "controls": false,
    "loop": false,
    "autoplay": false,
+   "startTime": -1,
+   "virtualClipDuration": -1,
    "preload": "auto",
    "theme": "path/to/videogular.css",
    "sources": [
@@ -116,6 +122,8 @@ angular.module("com.2fdevs.videogular")
             scope: {
                 vgTheme: "=?",
                 vgAutoPlay: "=?",
+                vgStartTime: "=?",
+                vgVirtualClipDuration: "=?",
                 vgPlaysInline: "=?",
                 vgNativeFullscreen: "=?",
                 vgClearMediaOnNavigate: "=?",